refactor(toggleButton): combine basic stories
