--- packages/design-system/src/stories/components/ToggleButton/SelectedColor.stories.tsx
@@ -1,174 +1,0 @@
-import React from "react";
-import {
-  Box,
-  Table,
-  TableHead,
-  TableRow,
-  TableBody,
-  TableCell,
-  Typography,
-} from "@mui/material";
-import { action } from "@storybook/addon-actions";
-
-import ToggleButton from "@/components/ToggleButton";
-
-import type { ComponentStory, ComponentMeta } from "@storybook/react";
-
-export default {
-  title: "Components/ToggleButton",
-  component: ToggleButton,
-  argTypes: {
-    value: {
-      control: false,
-      table: { type: { summary: "any" } },
-    },
-    selected: {
-      control: {
-        type: "boolean",
-      },
-      table: {
-        defaultValue: { summary: "false" },
-      },
-    },
-    selectedColor: {
-      control: false,
-      options: ["primary", "secondary"],
-      description:
-        "The color of the button when it is in an select state. It has nothing to do with active button",
-      table: {
-        defaultValue: { summary: "primary" },
-      },
-    },
-    color: {
-      control: {
-        type: "radio",
-      },
-      options: ["primary", "secondary"],
-      defaultValue: "primary",
-      description: "Button has three Kinds Contained, Ghost",
-      table: {
-        defaultValue: { summary: "primary" },
-      },
-    },
-    kind: {
-      control: {
-        type: "radio",
-      },
-      options: ["contained", "outlined", "ghost"],
-      defaultValue: "contained",
-      table: {
-        defaultValue: { summary: "contained" },
-      },
-    },
-    children: {
-      type: "string",
-      defaultValue: "Text",
-    },
-    disabled: {
-      control: {
-        type: "boolean",
-      },
-      defaultValue: false,
-      table: {
-        defaultValue: { summary: "false" },
-      },
-    },
-    size: {
-      control: {
-        type: "radio",
-      },
-      options: ["small", "medium", "large"],
-      defaultValue: "small",
-      table: {
-        defaultValue: { summary: "small" },
-      },
-    },
-    onClick: {
-      type: "function",
-      control: {
-        type: "select",
-      },
-      options: ["function", undefined],
-      mapping: {
-        function: action("onClick"),
-        undefined: undefined,
-      },
-      defaultValue: "function",
-      description:
-        "It is a callback function that is called when the button is clicked.",
-    },
-    onChange: {
-      type: "function",
-      control: false,
-      options: ["function", undefined],
-      mapping: {
-        function: action("onChange"),
-        undefined: undefined,
-      },
-      defaultValue: "function",
-    },
-  },
-  args: {
-    selected: true,
-    kind: "contained",
-  },
-  parameters: {
-    controls: {
-      include: [
-        "value",
-        "kind",
-        "selectedColor",
-        "onClick",
-        "children",
-        "disabled",
-        "size",
-        "selected",
-        "color",
-        "onChange",
-      ],
-    },
-    docs: {
-      description: {
-        component: `A Toggle Button can be used to group related options. please
-				see [Material-UI Toggle Button](https://mui.com/material-ui/react-toggle-button/).`,
-      },
-    },
-  },
-} as ComponentMeta<typeof ToggleButton>;
-
-const Template: ComponentStory<typeof ToggleButton> = (arg) => (
-  <Table sx={{ width: 330 }}>
-    <TableHead>
-      <TableRow>
-        <TableCell>
-          <Typography variant="body2_m">Selected color</Typography>
-        </TableCell>
-        <TableCell>
-          <Typography variant="body2_m">Primary</Typography>
-        </TableCell>
-        <TableCell>
-          <Typography variant="body2_m">Secondary</Typography>
-        </TableCell>
-      </TableRow>
-    </TableHead>
-    <TableBody>
-      <TableRow>
-        <TableCell>
-          <Typography variant="body2_m">Selected</Typography>
-        </TableCell>
-        <TableCell>
-          <ToggleButton {...arg} value="first">
-            {arg.children}
-          </ToggleButton>
-        </TableCell>
-        <TableCell>
-          <ToggleButton {...arg} selectedColor="secondary" value="second">
-            {arg.children}
-          </ToggleButton>
-        </TableCell>
-      </TableRow>
-    </TableBody>
-  </Table>
-);
-
-export const SelectedColor = Template.bind({});
